requirements.txt
torch~=2.1.1
wandb~=0.16.1
numpy~=1.24.2
pandas~=1.5.3
Pillow~=10.2.0
matplotlib~=3.7.1
scikit-learn~=1.3.2
torchvision~=0.16.1
tqdm~=4.65.0
typer~=0.9.0
python-dotenv~=1.0.0
